# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/persons/": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить информацию о субъекте доступа СКУД'",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Получение информации о субъекте доступа СКУД",
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса получения информации о субъекте доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.PersonData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Структура неудачного ответа выполнения запроса получения информации о субъекте доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Метод API, позволяющий авторизированному пользователю изменить информацию о субъекте доступа СКУД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Обновление информации о субъекте доступа СКУД",
                "parameters": [
                    {
                        "description": "Тело запроса обновления информации о субъекте доступа СКУД, формата 'application/json', в котором передается информация о субъекте доступа",
                        "name": "UpdatePersonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrserv.PersonData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса обновления информации о субъекте доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.PersonData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Структура неудачного ответа выполнения запроса обновления информации о субъекте доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Метод API, позволяющий авторизированному пользователю добавить нового субъекта доступа СКУД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Добавление субъекта доступа СКУД",
                "parameters": [
                    {
                        "description": "Тело запроса добавления нового субъекта доступа СКУД, формата 'application/json', в котором передается информация о субъекте доступа",
                        "name": "AddPersonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrserv.PersonData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса добавления субъекта доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.PersonData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Структура неудачного ответа выполнения запроса добавления субъекта доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Метод API, позволяющий авторизированному пользователю удалить субъект доступа СКУД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Удаление субъекта доступа СКУД",
                "parameters": [
                    {
                        "description": "Тело запроса удаления субъекта доступа СКУД, формата 'application/json', в котором передается информация о субъекте доступа",
                        "name": "DeletePersonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/integrserv.PersonData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса удаления субъекта доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/integrserv.PersonData"
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Структура неудачного ответа выполнения запроса удаления субъекта доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/persons/activity/dayly": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить информацию о статистике посещаемости субъектом доступа за конкретный день",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Получение информации о статистике посещаемости субъектом доступа за день СКУД",
                "parameters": [
                    {
                        "type": "string",
                        "default": "2024-03-02T15:04:05-07:00",
                        "description": "День посещения",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1417,
                        "description": "Идентификатор субъекта доступа СКУД",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса получения информации о статистике посещаемости субъектом доступа СКУД за конкретный день",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/resp.Action"
                                            }
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Структура неудачного ответа выполнения запроса получения информации о статистике посещаемости субъектом доступа СКУД за конкретный день",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/persons/activity/monthly": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить информацию о статистике посещаемости субъектом доступа за конкретный месяц",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Получение информации о статистике посещаемости субъектом доступа за месяц СКУД",
                "parameters": [
                    {
                        "type": "string",
                        "default": "2024-03-02T15:04:05-07:00",
                        "description": "Месяц на который нужно расчитать статистику",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1417,
                        "description": "Идентификатор субъекта доступа СКУД",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса получения информации о статистике посещаемости субъектом доступа СКУД за конкретный месяц",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/resp.Activity"
                                            }
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Структура неудачного ответа выполнения запроса получения информации о статистике посещаемости субъектом доступа СКУД за конкретный месяц",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/persons/departments": {
            "get": {
                "description": "Метод API, позволяющий авторизированному пользователю получить информацию о группах субъектов доступа СКУД",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Получение информации о группах субъектов доступа СКУД",
                "responses": {
                    "200": {
                        "description": "Структура успешного ответа выполнения запроса получения информации о группах субъектов доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/integrserv.Department"
                                            }
                                        },
                                        "error": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Структура неудачного ответа выполнения запроса получения информации о группах субъектов доступа СКУД",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Body"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "integrserv.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "integrserv.PersonData": {
            "type": "object",
            "properties": {
                "departmentId": {
                    "type": "integer"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "middleName": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tabNum": {
                    "type": "string"
                }
            }
        },
...
        "resp.Action": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "resp.Activity": {
            "type": "object",
            "properties": {
                "coming": {
                    "type": "integer"
                },
                "leaving": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "response.Body": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {}
            }
        }
    }
}